Refined Card Design
Colours for the whole app will have to reviewed in polish-staging
