+++ src/renderer/Modules/Home/home.scss
@@ -1,0 +1,5 @@
+@use '../../style.scss';
+
+.r-home {
+  display: flex;
+}
