+++ src/renderer/Modules/Home/home.tsx
@@ -1,23 +1,24 @@
-import React, { useRef, useState } from 'react';
+import React, { useRef } from 'react';
 import path from 'path'
-import { ipcRenderer } from 'electron'
-import { Sound, InputText, Button } from "../../Components"
+import { Sound, Card } from "../../Components"
+import { Login } from "./login"
+
+import "./home.scss"
 
 export const Home = () => {
   let radioWaveMonster = useRef<any>(null);
-  const [ pass, setPass ] = useState<string>('')
-
-  const submit = async () => {
-    ipcRenderer.invoke('ssh-open', {password:pass})
-  }
 
 	return (
 	<div className="r-home">
-    <i className="fa fa-hard-hat"/>
-    This page is under construction
+    <Login />
 
-    <InputText type="password" value={pass} onChange={(e) => setPass(e.target.value)} />
-    <Button label="Submit" onClick={submit} />
+    <Card>
+      <Sound ref={radioWaveMonster} soundFile="assets/RODAN RODAN RADIO WAVE MONSTER.mp3" />
+      <img src="assets/icon.png" style={{width:"230px", height:"230px"}}
+              onClick={() => {
+                radioWaveMonster.current!.play()
+              }}/>
+    </Card>
   </div>
 	)
 }
