--- src/renderer/Modules/login.tsx
@@ -1,12 +1,0 @@
-import React, { useState } from 'react'
-import { ipcRenderer } from 'electron'
-import { InputText, Button } from '../Components'
-
-export function Login(props:{}) {
-
-  return (
-    <div>
-      TESTING LOGIN
-    </div>
-  )
-}
